PR diff LHS commit is reported as missing Fixes #3941
