+++ src/view/compareChangesTreeDataProvider.ts
@@ -3,15 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as buffer from 'buffer';
 import * as vscode from 'vscode';
 import { Repository } from '../api/api';
 import { getGitChangeType } from '../common/diffHunk';
 import Logger from '../common/logger';
-import { fromGitHubURI, Schemes } from '../common/uri';
+import { Schemes } from '../common/uri';
 import { FolderRepositoryManager } from '../github/folderRepositoryManager';
 import { GitHubRepository } from '../github/githubRepository';
-import { ReadonlyFileSystemProvider } from './readonlyFileSystemProvider';
+import { GitHubContentProvider } from './gitHubContentProvider';
 import { GitHubFileChangeNode } from './treeNodes/fileChangeNode';
 import { TreeNode } from './treeNodes/treeNode';
 
@@ -174,44 +173,4 @@ export class CompareChangesTreeProvider implements vscode.TreeDataProvider<TreeN
 	}
 }
 
-/**
- * Provides file contents for documents with GITHUB_FILE_SCHEME (githubpr) scheme. Contents are fetched from GitHub based on
- * information in the document's query string.
- */
-class GitHubContentProvider extends ReadonlyFileSystemProvider {
-	constructor(public gitHubRepository: GitHubRepository) {
-		super();
-	}
 
-	async readFile(uri: any): Promise<Uint8Array> {
-		const params = fromGitHubURI(uri);
-		if (!params || params.isEmpty) {
-			return new TextEncoder().encode('');
-		}
-
-		const { octokit, remote } = await this.gitHubRepository.ensure();
-		let fileContent: { data: { content: string; encoding: string; sha: string } } = (await octokit.call(octokit.api.repos.getContent,
-			{
-				owner: remote.owner,
-				repo: remote.repositoryName,
-				path: params.fileName,
-				ref: params.branch,
-			},
-		)) as any;
-		let contents = fileContent.data.content ?? '';
-
-		// Empty contents and 'none' encoding indcates that the file has been truncated and we should get the blob.
-		if (contents === '' && fileContent.data.encoding === 'none') {
-			const fileSha = fileContent.data.sha;
-			fileContent = await octokit.call(octokit.api.git.getBlob, {
-				owner: remote.owner,
-				repo: remote.repositoryName,
-				file_sha: fileSha,
-			});
-			contents = fileContent.data.content;
-		}
-
-		const buff = buffer.Buffer.from(contents, (fileContent.data as any).encoding);
-		return buff;
-	}
-}
